Moved Priority Queue Functions
Moved from api\api_factory.py to api\backlog_controller.py
